Projeto Trello/trelloso_fastapi/requirements.txt
# General
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
pydantic_settings

# PostgreSQL
psycopg2-binary

# Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart
